refactor: move the shared preference extensions into new package
